Create cal.html
This code is about a Calculators require an interactive interface tc perform basic functions such as addition, subtraction, division, and multiplication. It will need a display screen t display the user's input and give the results. The grid syst in CSS is for the alignments of buttons on the calculator. The additional tools i build a fully functional calculator include eventListeners, if-else statements, operators, loops, and so on.